major change to document model
Retain operations no longer exist and Skip is replaced by Delete

Inserts and Deletes now refer to an "id" and a character offset in the
entire replacement text associated with the id (e.g. for history
blocks, it is the concatenation of the text of all of the
replacements).

When inserted text is deleted, a Delete operation is added containing
the deleted text. This provides information required to prevent merges
from duplicating deletes and inserts.

The id of original document text, so a "Retain" operation is now an
Insert with id "" and a "Skip" operation is now a Delete with id "".